[Bifrost] Move Record to restate-types
Another mechanical change to make this type available to log-server in future PRs
