+++ crates/bifrost/src/providers/memory_loglet.rs
@@ -23,14 +23,13 @@ use tracing::{debug, info};
 
 use restate_core::ShutdownError;
 use restate_types::logs::metadata::{LogletParams, ProviderKind, SegmentIndex};
-use restate_types::logs::{KeyFilter, LogId, LogletOffset, MatchKeyQuery, SequenceNumber};
+use restate_types::logs::{KeyFilter, LogId, LogletOffset, MatchKeyQuery, Record, SequenceNumber};
 
 use crate::loglet::util::TailOffsetWatch;
 use crate::loglet::{
     Loglet, LogletCommit, LogletProvider, LogletProviderFactory, LogletReadStream, OperationError,
     SendableLogletReadStream,
 };
-use crate::Record;
 use crate::Result;
 use crate::{LogEntry, TailState};
